requirements.txt
PySide6>=6.6.0
Pillow>=10.0.0
pytest>=7.0.0
pytest-qt>=4.2.0
